requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pillow==10.1.0
python-multipart==0.0.6
